fix. BookingService 분리
우선 내부의 기능별로 분리함
